switching to ground_truth_models structure
