+++ predicators/ground_truth_models/cover/__init__.py
@@ -1,0 +1,3 @@
+"""Ground-truth models for cover environment and variants."""
+
+from .options import GroundTruthOptionFactory
