+++ predicators/nsrt_learning/strips_learning/oracle_learner.py
@@ -4,8 +4,8 @@ import logging
 from typing import List
 
 from predicators.envs import get_or_create_env
+from predicators.ground_truth_models import get_gt_options
 from predicators.ground_truth_nsrts import get_gt_nsrts
-from predicators.ground_truth_options import get_gt_options
 from predicators.nsrt_learning.strips_learning import BaseSTRIPSLearner
 from predicators.settings import CFG
 from predicators.structs import PNAD, Datastore, DummyOption
