+++ predicators/option_model.py
@@ -13,7 +13,7 @@ import numpy as np
 
 from predicators import utils
 from predicators.envs import BaseEnv, create_new_env
-from predicators.ground_truth_options import get_gt_options
+from predicators.ground_truth_models import get_gt_options
 from predicators.settings import CFG
 from predicators.structs import DefaultState, State, _Option
