+++ scripts/evaluate_interactive_approach_classifiers.py
@@ -13,7 +13,7 @@ from predicators.approaches.interactive_learning_approach import \
     InteractiveLearningApproach
 from predicators.envs import BaseEnv, create_new_env
 from predicators.envs.cover import CoverEnv
-from predicators.ground_truth_options import get_gt_options
+from predicators.ground_truth_models import get_gt_options
 from predicators.settings import CFG
 from predicators.structs import Object, Predicate, State, Task
 from scripts.analyze_results_directory import get_df_for_entry
