+++ scripts/skeleton_score_analysis.py
@@ -18,7 +18,7 @@ from numpy.typing import NDArray
 from predicators import utils
 from predicators.datasets import create_dataset
 from predicators.envs import create_new_env
-from predicators.ground_truth_options import get_gt_options
+from predicators.ground_truth_models import get_gt_options
 from predicators.nsrt_learning.segmentation import segment_trajectory
 from predicators.nsrt_learning.strips_learning import learn_strips_operators
 from predicators.planning import PlanningFailure, PlanningTimeout, task_plan, \
